requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
dj-database-url==0.5.0
Django==3.2
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.11.2
django-embed-video==1.4.0
django-heroku==0.3.1
django-mathfilters==1.0.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.6.0
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
zipp==3.4.1
